dedicated team chat functional, bug with create team
