+++ client/components/team/TeamView.js
@@ -1,6 +1,5 @@
 import React from 'react';
-import fire from '../../fire';
-
+import { Link } from 'react-router-dom';
 /*
 * https://www.codementor.io/yurio/all-you-need-is-react-firebase-4v7g9p4kf
 */
@@ -8,37 +7,6 @@ import fire from '../../fire';
 export default class TeamView extends React.Component {
   constructor(props) {
     super(props);
-    this.state = {
-      messages: [],
-      newMessage: ''
-    };
-  }
-
-  componentWillMount() {
-    // Ref for messages in firebase
-    let firebase = fire.database();
-    let teamChat = firebase.ref('team_' + this.props.teamId + '/messages')
-                    .orderByKey()
-                    .limitToLast(25);
-
-    teamChat.on('child_added', snapshot => {
-      // Update state when new message added to Firebase db
-      let message = { text: snapshot.val(), id: snapshot.key };
-      this.setState({ messages: [message].concat(this.state.messages) });
-    });
-  }
-
-  addMessage(e) {
-    e.preventDefault();
-    let firebase = fire.database();
-    let teamChat = firebase.ref('team_' + this.props.teamId + '/messages');
-    // Send new message to the db
-    teamChat.push(this.state.newMessage);
-    this.setState({ newMessage: '' });
-  }
-
-  handleChange(e) {
-    this.setState({ newMessage: e.target.value });
   }
 
   render() {
@@ -105,25 +73,8 @@ export default class TeamView extends React.Component {
           {teamList(this.props.activeTeam)}
         </div>
 
-        <form
-          className='message-view-form'
-          onSubmit={e => this.addMessage(e)}
-          >
-          <label htmlFor='new-message'>Send a message:</label>
-          <input
-            type='text'
-            value={this.state.newMessage}
-            onChange={e => this.handleChange(e)}
-          />
-          <button type='submit'>Send</button>
-          <ul className='messages-list'>
-            {
-              // Render the messages
-              this.state.messages.map(m => <li key={m.id}>{m.text}</li>)
-            }
-          </ul>
+        <Link to={`/chat/team/${this.props.teamId}`}> Team Chat </Link>
 
-        </form>
       </div>
     );
   }
